APLICACION/requirements.txt
beautifulsoup4==4.9.3
pandas==1.2.4
requests==2.25.1
SQLAlchemy==1.4.15
python-dotenv==0.17.1
pyodbc==4.0.30
